Support for schemas and some unit test corrections
Added Table factory method taking a schema parameter.
Removed the Table factory method taking the alias, as the same thing can be achieved with the As extension method. The Column and Table factory methods are now more consistent with each other, as they both have overloads where a prefix can be provided.
Added missing "p0" parameter in syntax tests for Atan2 function for some of the translators.